requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncpg>=0.28.0
aiohttp>=3.9.0
httpx>=0.25.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-docx>=1.1.0
langchain-community>=0.0.0